refactor: refactor MncSketch.h into MncSketch-Propagate,Estimate and Build files
